Split init modul into separate modules.
